Refactor to separate component
