+++ frontend/src/scenes/experiments/ExperimentMetricForm.tsx
@@ -1,19 +1,14 @@
 import { DataWarehousePopoverField } from 'lib/components/TaxonomicFilter/types'
 import { LemonLabel } from 'lib/lemon-ui/LemonLabel'
 import { LemonRadio } from 'lib/lemon-ui/LemonRadio'
-import { LemonSelect } from 'lib/lemon-ui/LemonSelect'
 import { ActionFilter } from 'scenes/insights/filters/ActionFilter/ActionFilter'
 
 import { Query } from '~/queries/Query/Query'
-import {
-    ExperimentMetric,
-    ExperimentMetricType,
-    isExperimentFunnelMetric,
-    NodeKind,
-} from '~/queries/schema/schema-general'
-import { FilterType, StepOrderValue } from '~/types'
+import { ExperimentMetric, ExperimentMetricType, NodeKind } from '~/queries/schema/schema-general'
+import { FilterType } from '~/types'
 
 import { ExperimentMetricConversionWindowFilter } from './ExperimentMetricConversionWindowFilter'
+import { ExperimentMetricFunnelOrderFilter } from './ExperimentMetricFunnelOrderFilter'
 import { ExperimentMetricOutlierHandling } from './ExperimentMetricOutlierHandling'
 import { commonActionFilterProps } from './Metrics/Selectors'
 import {
@@ -44,33 +39,6 @@ const dataWarehousePopoverFields: DataWarehousePopoverField[] = [
     },
 ]
 
-const funnelOrderOptions = [
-    {
-        label: 'Sequential',
-        value: StepOrderValue.ORDERED,
-    },
-    {
-        label: 'Any order',
-        value: StepOrderValue.UNORDERED,
-    },
-]
-
-function StepOrderInfo(): JSX.Element {
-    return (
-        <ul className="list-disc pl-4">
-            <li>
-                <b>Sequential</b> - Step B must happen after Step A, but any number events can happen between A and B.
-            </li>
-            <li>
-                <b>Strict order</b> - Step B must happen directly after Step A without any events in between.
-            </li>
-            <li>
-                <b>Any order</b> - Steps can be completed in any sequence.
-            </li>
-        </ul>
-    )
-}
-
 export function ExperimentMetricForm({
     metric,
     handleSetMetric,
@@ -95,15 +63,6 @@ export function ExperimentMetricForm({
 
     const handleMetricTypeChange = (newMetricType: ExperimentMetricType): void => {
         handleSetMetric(getDefaultExperimentMetric(newMetricType))
-    }
-
-    const handleFunnelOrderTypeChange = (funnelOrderType: StepOrderValue): void => {
-        if (isExperimentFunnelMetric(metric)) {
-            handleSetMetric({
-                ...metric,
-                funnel_order_type: funnelOrderType,
-            })
-        }
     }
 
     const radioOptions = [
@@ -186,20 +145,7 @@ export function ExperimentMetricForm({
                     />
                 )}
             </div>
-            {isExperimentFunnelMetric(metric) && (
-                <div>
-                    <LemonLabel className="mb-1" info={<StepOrderInfo />}>
-                        Step order
-                    </LemonLabel>
-                    <LemonSelect
-                        data-attr="experiment-funnel-order-filter"
-                        value={metric.funnel_order_type || StepOrderValue.ORDERED}
-                        onChange={handleFunnelOrderTypeChange}
-                        dropdownMatchSelectWidth={false}
-                        options={funnelOrderOptions}
-                    />
-                </div>
-            )}
+            <ExperimentMetricFunnelOrderFilter metric={metric} handleSetMetric={handleSetMetric} />
             <ExperimentMetricConversionWindowFilter metric={metric} handleSetMetric={handleSetMetric} />
             {metric.metric_type === ExperimentMetricType.MEAN && (
                 <ExperimentMetricOutlierHandling metric={metric} handleSetMetric={handleSetMetric} />
